Refactor attribute utils
